Allow PerfTests to control what ITraceManager is used
