+++ src/Test/Perf/Utilities/Perf.Utilities.csproj
@@ -45,10 +45,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ITraceManager.cs" />
     <Compile Include="Logger.cs" />
     <Compile Include="PerfTestBase.cs" />
     <Compile Include="RelativeDirectory.cs" />
+    <Compile Include="ScenarioGenerator.cs" />
     <Compile Include="TestUtilities.cs" />
+    <Compile Include="TraceManager.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Properties\" />
